+++ src/components/ExternalLink.tsx
@@ -1,10 +1,10 @@
 import styled from 'styled-components';
 
-const A = styled(({ children, ...props }: any) => <a target="_blank" rel="noopener noreferrer" {...props} >{children}</a>)`
+const ExternalLink = styled(({ children, ...props }: any) => <a target="_blank" rel="noopener noreferrer" {...props} >{children}</a>)`
   transition: color 0.1s;
   color: var(--blue);
 
   &:hover { color: unset; }
 `;
 
-export default A;
+export default ExternalLink;